+++ wechatpy/work/client/__init__.py
@@ -27,7 +27,7 @@ class WeChatClient(BaseWeChatClient):
     shakearound = api.WeChatShakeAround()
     tag = api.WeChatTag()
     user = api.WeChatUser()
-    OA = api.WeChatOA()
+    oa = api.WeChatOA()
 
     def __init__(self, corp_id, secret, access_token=None,
                  session=None, timeout=None, auto_retry=True):
