+++ wechatpy/work/client/api/oauth.py
@@ -1,4 +1,4 @@
-# encoding: utf-8
+# -*- coding: utf-8 -*-
 
 from urllib.parse import quote
